Move reloader to run-support project
