+++ framework/src/sbt-plugin/src/main/scala/PlayRun.scala
@@ -20,14 +20,13 @@ import java.util.jar.JarFile
 import com.typesafe.sbt.SbtNativePackager._
 import com.typesafe.sbt.packager.Keys._
 import com.typesafe.sbt.web.SbtWeb.autoImport._
-import play.runsupport.{ PlayWatchService, AssetsClassLoader }
+import play.runsupport.{ AssetsClassLoader, PlayWatchService, Reloader }
 import play.sbtplugin.run._
 
 /**
  * Provides mechanisms for running a Play application in SBT
  */
 trait PlayRun extends PlayInternalKeys {
-  this: PlayReloader =>
 
   /**
    * Configuration for the Play docs application's dependencies. Used to build a classloader for
@@ -35,49 +34,8 @@ trait PlayRun extends PlayInternalKeys {
    */
   val DocsApplication = config("docs").hide
 
-  // Regex to match Java System Properties of the format -Dfoo=bar
-  private val SystemProperty = "-D([^=]+)=(.*)".r
-
-  /**
-   * Take all the options in javaOptions of the format "-Dfoo=bar" and return them as a Seq of key value pairs of the format ("foo" -> "bar")
-   */
-  private def extractSystemProperties(javaOptions: Seq[String]): Seq[(String, String)] = {
-    javaOptions.collect { case SystemProperty(key, value) => key -> value }
-  }
-
-  private def parsePort(portString: String): Int = {
-    try {
-      Integer.parseInt(portString)
-    } catch {
-      case e: NumberFormatException => sys.error("Invalid port argument: " + portString)
-    }
-  }
-
-  private def filterArgs(args: Seq[String], defaultHttpPort: Int): (Seq[(String, String)], Option[Int], Option[Int]) = {
-    val (properties, others) = args.span(_.startsWith("-D"))
-
-    val javaProperties = properties.map(_.drop(2).split('=')).map(a => a(0) -> a(1)).toSeq
-
-    // collect arguments plus config file property if present
-    val httpPort = Option(System.getProperty("http.port"))
-    val httpsPort = Option(System.getProperty("https.port"))
-
-    //port can be defined as a numeric argument or as disabled, -Dhttp.port argument or a generic sys property
-    val maybePort = others.headOption.orElse(javaProperties.toMap.get("http.port")).orElse(httpPort)
-    val maybeHttpsPort = javaProperties.toMap.get("https.port").orElse(httpsPort).map(parsePort)
-    if (maybePort.exists(_ == "disabled")) (javaProperties, Option.empty[Int], maybeHttpsPort)
-    else (javaProperties, maybePort.map(parsePort).orElse(Some(defaultHttpPort)), maybeHttpsPort)
-  }
-
-  val createURLClassLoader: ClassLoaderCreator = (name, urls, parent) => new java.net.URLClassLoader(urls, parent) {
-    override def toString = name + "{" + getURLs.map(_.toString).mkString(", ") + "}"
-  }
-
-  val createDelegatedResourcesClassLoader: ClassLoaderCreator = (name, urls, parent) => new java.net.URLClassLoader(urls, parent) {
-    require(parent ne null)
-    override def getResources(name: String): java.util.Enumeration[java.net.URL] = getParent.getResources(name)
-    override def toString = name + "{" + getURLs.map(_.toString).mkString(", ") + "}"
-  }
+  val createURLClassLoader: ClassLoaderCreator = Reloader.createURLClassLoader
+  val createDelegatedResourcesClassLoader: ClassLoaderCreator = Reloader.createDelegatedResourcesClassLoader
 
   val playDefaultRunTask = playRunTask(playRunHooks, playDependencyClasspath, playDependencyClassLoader,
     playReloaderClasspath, playReloaderClassLoader, playAssetsClassLoader)
@@ -100,22 +58,36 @@ trait PlayRun extends PlayInternalKeys {
     val state = Keys.state.value
     val interaction = playInteractionMode.value
 
-    lazy val devModeServer = startDevMode(
-      state,
+    val reloadCompile = () => PlayReload.compile(
+      () => Project.runTask(playReload, state).map(_._2).get,
+      () => Project.runTask(reloaderClasspath, state).map(_._2).get,
+      () => Project.runTask(streamsManager, state).map(_._2).get.toEither.right.toOption
+    )
+
+    val runSbtTask: String => AnyRef = (task: String) => {
+      val parser = Act.scopedKeyParser(state)
+      val Right(sk) = complete.DefaultParsers.result(parser, task)
+      val result = Project.runTask(sk.asInstanceOf[Def.ScopedKey[Task[AnyRef]]], state).map(_._2)
+      result.flatMap(_.toEither.right.toOption).orNull
+    }
+
+    lazy val devModeServer = Reloader.startDevMode(
       runHooks.value,
       (javaOptions in Runtime).value,
-      dependencyClasspath.value,
+      dependencyClasspath.value.files,
       dependencyClassLoader.value,
-      reloaderClasspath,
+      reloadCompile,
       reloaderClassLoader.value,
       assetsClassLoader.value,
       playCommonClassloader.value,
       playMonitoredFiles.value,
       playWatchService.value,
-      (managedClasspath in DocsApplication).value,
-      interaction,
+      (managedClasspath in DocsApplication).value.files,
       playDefaultPort.value,
-      args
+      baseDirectory.value,
+      devSettings.value,
+      args,
+      runSbtTask
     )
 
     interaction match {
@@ -201,162 +173,6 @@ trait PlayRun extends PlayInternalKeys {
     }
   }
 
-  /**
-   * Play dev server
-   */
-  private trait PlayDevServer extends Closeable {
-    val buildLink: BuildLink
-  }
-
-  /**
-   * Start the Play server in dev mode
-   *
-   * @return A closeable that can be closed to stop the server
-   */
-  private def startDevMode(state: State, runHooks: Seq[play.PlayRunHook], javaOptions: Seq[String],
-    dependencyClasspath: Classpath, dependencyClassLoader: ClassLoaderCreator,
-    reloaderClasspathTask: TaskKey[Classpath], reloaderClassLoader: ClassLoaderCreator,
-    assetsClassLoader: ClassLoader => ClassLoader, commonClassLoader: ClassLoader,
-    monitoredFiles: Seq[String], playWatchService: PlayWatchService,
-    docsClasspath: Classpath, interaction: PlayInteractionMode, defaultHttpPort: Int,
-    args: Seq[String]): PlayDevServer = {
-
-    val (properties, httpPort, httpsPort) = filterArgs(args, defaultHttpPort = defaultHttpPort)
-    val systemProperties = extractSystemProperties(javaOptions)
-
-    require(httpPort.isDefined || httpsPort.isDefined, "You have to specify https.port when http.port is disabled")
-
-    // Set Java properties
-    (properties ++ systemProperties).foreach {
-      case (key, value) => System.setProperty(key, value)
-    }
-
-    println()
-
-    /*
-     * We need to do a bit of classloader magic to run the Play application.
-     *
-     * There are seven classloaders:
-     *
-     * 1. buildLoader, the classloader of sbt and the Play sbt plugin.
-     * 2. commonLoader, a classloader that persists across calls to run.
-     *    This classloader is stored inside the
-     *    PlayInternalKeys.playCommonClassloader task. This classloader will
-     *    load the classes for the H2 database if it finds them in the user's
-     *    classpath. This allows H2's in-memory database state to survive across
-     *    calls to run.
-     * 3. delegatingLoader, a special classloader that overrides class loading
-     *    to delegate shared classes for build link to the buildLoader, and accesses
-     *    the reloader.currentApplicationClassLoader for resource loading to
-     *    make user resources available to dependency classes.
-     *    Has the commonLoader as its parent.
-     * 4. applicationLoader, contains the application dependencies. Has the
-     *    delegatingLoader as its parent. Classes from the commonLoader and
-     *    the delegatingLoader are checked for loading first.
-     * 5. docsLoader, the classloader for the special play-docs application
-     *    that is used to serve documentation when running in development mode.
-     *    Has the applicationLoader as its parent for Play dependencies and
-     *    delegation to the shared sbt doc link classes.
-     * 6. playAssetsClassLoader, serves assets from all projects, prefixed as
-     *    configured.  It does no caching, and doesn't need to be reloaded each
-     *    time the assets are rebuilt.
-     * 7. reloader.currentApplicationClassLoader, contains the user classes
-     *    and resources. Has applicationLoader as its parent, where the
-     *    application dependencies are found, and which will delegate through
-     *    to the buildLoader via the delegatingLoader for the shared link.
-     *    Resources are actually loaded by the delegatingLoader, where they
-     *    are available to both the reloader and the applicationLoader.
-     *    This classloader is recreated on reload. See PlayReloader.
-     *
-     * Someone working on this code in the future might want to tidy things up
-     * by splitting some of the custom logic out of the URLClassLoaders and into
-     * their own simpler ClassLoader implementations. The curious cycle between
-     * applicationLoader and reloader.currentApplicationClassLoader could also
-     * use some attention.
-     */
-
-    // Get the URLs for the resources in a classpath
-    def urls(cp: Classpath): Array[URL] = cp.map(_.data.toURI.toURL).toArray
-
-    val buildLoader = this.getClass.getClassLoader
-
-    /**
-     * ClassLoader that delegates loading of shared build link classes to the
-     * buildLoader. Also accesses the reloader resources to make these available
-     * to the applicationLoader, creating a full circle for resource loading.
-     */
-    lazy val delegatingLoader: ClassLoader = new DelegatingClassLoader(commonClassLoader, buildLoader, new ApplicationClassLoaderProvider {
-      def get: ClassLoader = { reloader.getClassLoader.orNull }
-    })
-
-    lazy val applicationLoader = dependencyClassLoader("PlayDependencyClassLoader", urls(dependencyClasspath), delegatingLoader)
-    lazy val assetsLoader = assetsClassLoader(applicationLoader)
-
-    lazy val reloader = newReloader(state, playReload, reloaderClassLoader, reloaderClasspathTask, assetsLoader,
-      monitoredFiles, playWatchService)
-
-    try {
-      // Now we're about to start, let's call the hooks:
-      runHooks.run(_.beforeStarted())
-
-      // Get a handler for the documentation. The documentation content lives in play/docs/content
-      // within the play-docs JAR.
-      val docsLoader = new URLClassLoader(urls(docsClasspath), applicationLoader)
-      val docsJarFile = {
-        val f = docsClasspath.map(_.data).filter(_.getName.startsWith("play-docs")).head
-        new JarFile(f)
-      }
-      val buildDocHandler = {
-        val docHandlerFactoryClass = docsLoader.loadClass("play.docs.BuildDocHandlerFactory")
-        val factoryMethod = docHandlerFactoryClass.getMethod("fromJar", classOf[JarFile], classOf[String])
-        factoryMethod.invoke(null, docsJarFile, "play/docs/content").asInstanceOf[BuildDocHandler]
-      }
-
-      val server = {
-        val mainClass = applicationLoader.loadClass("play.core.server.NettyServer")
-        if (httpPort.isDefined) {
-          val mainDev = mainClass.getMethod("mainDevHttpMode", classOf[BuildLink], classOf[BuildDocHandler], classOf[Int])
-          mainDev.invoke(null, reloader, buildDocHandler, httpPort.get: java.lang.Integer).asInstanceOf[play.core.server.ServerWithStop]
-        } else {
-          val mainDev = mainClass.getMethod("mainDevOnlyHttpsMode", classOf[BuildLink], classOf[BuildDocHandler], classOf[Int])
-          mainDev.invoke(null, reloader, buildDocHandler, httpsPort.get: java.lang.Integer).asInstanceOf[play.core.server.ServerWithStop]
-        }
-      }
-
-      // Notify hooks
-      runHooks.run(_.afterStarted(server.mainAddress))
-
-      new PlayDevServer {
-        val buildLink = reloader
-
-        def close() = {
-          server.stop()
-          docsJarFile.close()
-          reloader.close()
-
-          // Notify hooks
-          runHooks.run(_.afterStopped())
-
-          // Remove Java properties
-          properties.foreach {
-            case (key, _) => System.clearProperty(key)
-          }
-        }
-      }
-    } catch {
-      case e: Throwable =>
-        // Let hooks clean up
-        runHooks.foreach { hook =>
-          try {
-            hook.onError()
-          } catch {
-            case e: Throwable => // Swallow any exceptions so that all `onError`s get called.
-          }
-        }
-        throw e
-    }
-  }
-
   val playPrefixAndAssetsSetting = playPrefixAndAssets := {
     assetsPrefix.value -> (WebKeys.public in Assets).value
   }
@@ -405,7 +221,7 @@ trait PlayRun extends PlayInternalKeys {
 
     val interaction = extracted.get(playInteractionMode)
     // Parse HTTP port argument
-    val (properties, httpPort, httpsPort) = filterArgs(args, defaultHttpPort = extracted.get(playDefaultPort))
+    val (properties, httpPort, httpsPort) = Reloader.filterArgs(args, defaultHttpPort = extracted.get(playDefaultPort))
     require(httpPort.isDefined || httpsPort.isDefined, "You have to specify https.port when http.port is disabled")
 
     Project.runTask(stage, state).get._2.toEither match {
