+++ framework/src/sbt-plugin/src/main/scala/play/Project.scala
@@ -17,7 +17,6 @@ import play.twirl.sbt.SbtTwirl
 object Play
   extends AutoPlugin
   with PlayExceptions
-  with PlayReloader
   with PlayCommands
   with PlayRun
   with play.PlaySettings
