+++ framework/src/sbt-plugin/src/test/scala/play/PlayRunHookSpec.scala
@@ -14,7 +14,7 @@ object PlayRunHookSpec extends Specification {
 
     "provide implicit `run` which passes every hook to a provided function" in {
       val hooks = Seq.fill(3)(new PlayRunHook {})
-      val executedHooks: HashMap[PlayRunHook, Boolean] = HashMap.empty
+      val executedHooks: HashMap[play.runsupport.RunHook, Boolean] = HashMap.empty
 
       hooks.run(hook => executedHooks += ((hook, true)))
 
@@ -22,7 +22,7 @@ object PlayRunHookSpec extends Specification {
     }
 
     "re-throw an exception on single hook failure" in {
-      val executedHooks: HashMap[PlayRunHook, Boolean] = HashMap.empty
+      val executedHooks: HashMap[play.runsupport.RunHook, Boolean] = HashMap.empty
       class HookMockException extends Throwable
 
       val hooks = Seq.fill(3)(new PlayRunHook {
@@ -37,7 +37,7 @@ object PlayRunHookSpec extends Specification {
     }
 
     "combine several thrown exceptions into a RunHookCompositeThrowable" in {
-      val executedHooks: HashMap[PlayRunHook, Boolean] = HashMap.empty
+      val executedHooks: HashMap[play.runsupport.RunHook, Boolean] = HashMap.empty
       class HookFirstMockException extends Throwable
       class HookSecondMockException extends Throwable
 
@@ -57,7 +57,7 @@ object PlayRunHookSpec extends Specification {
 
       val hooks = firstFailure +: dummyHooks :+ lastFailure
 
-      hooks.run(_.beforeStarted()) must throwAn[RunHookCompositeThrowable].like {
+      hooks.run(_.beforeStarted()) must throwA[play.runsupport.RunHookCompositeThrowable].like {
         case e: Throwable =>
           e.getMessage must contain("HookFirstMockException")
           e.getMessage must contain("HookSecondMockException")
